Build Standalone Configuration
Refactor application and image build system to fully support 3 variants of the LunaCam image:

* Camera-Only
* Portal-Only
* Full (portal and camera)

+++ tools/scripts/build-image.ps1
@@ -18,12 +18,6 @@ $pigenDir = Join-Path (Join-Path $repoDir tools) pi-gen
 $pigenBuildDir = Join-Path $buildDir pi-gen
 $rustBuildDir = "$buildDir/target/arm-unknown-linux-gnueabihf/release"
 
-# Build prerequisites
-& "$PSScriptRoot/build-lcsvc.ps1" -Variant $Variant
-if (($Variant -eq "Full") -or ($Variant -eq "PortalOnly")) {
-    & "$PSScriptRoot/build-css.ps1"
-}
-
 if (!(Test-Path $pigenBuildDir)) {
     Write-Host "cloning pi-gen"
     git clone --depth 1 https://github.com/RPi-Distro/pi-gen $pigenBuildDir
@@ -40,52 +34,65 @@ function prepareStage {
     Remove-Item -Recurse -Force $pigenBuildDir/$name -ErrorAction Ignore
     Copy-Item -Recurse $pigenDir/$name $pigenBuildDir/$name
     Copy-Item $pigenDir/prerun.sh $pigenBuildDir/$name/prerun.sh
+
+    # Ensure all copied scripts are executable
+    Get-ChildItem -Recurse $pigenBuildDir/$name |
+        Where-Object Name -Match "\.sh$" |
+        Select-Object -ExpandProperty FullName |
+        xargs chmod +x
 }
 
 prepareStage "common"
+$commonDir = "$pigenBuildDir/common"
+
+& "$PSScriptRoot/build-lcsvc.ps1" -Variant $Variant
+Copy-Item $rustBuildDir/lcsvc $commonDir/03-lcsvc/files/lcsvc
+
+$templateDir = "$commonDir/03-lcsvc/files/templates"
+$null = New-Item -Type Directory $templateDir
+Copy-Item -Recurse $repoDir/templates/* $templateDir
+
 Copy-Item $pigenDir/config.sh $pigenBuildDir/config
 
-switch ($Variant) {
-    "Full" {
-        throw "full image variant not yet supported"
-    }
-    "PortalOnly" {
-        prepareStage "portal"
+$stageList = "stage0 stage1 stage2 common"
 
-        $portalDir = "$pigenBuildDir/portal"
+if ($Variant -ne "CameraOnly") {
 
-        Copy-Item $rustBuildDir/lcsvc $portalDir/01-portal/files/lcportal
+    prepareStage "portal"
+    $portalDir = "$pigenBuildDir/portal"
 
-        $staticDir = "$portalDir/01-portal/files/static"
-        Remove-Item -Recurse -Force $staticDir -ErrorAction Ignore
-        $null = New-Item -Type Directory $staticDir
-        Copy-Item $buildDir/css $staticDir/css
-        Copy-Item $repoDir/client/js $staticDir/js
+    & "$PSScriptRoot/build-css.ps1"
 
-        $templateDir = "$portalDir/01-portal/files/templates"
-        Remove-Item -Recurse -Force $templateDir -ErrorAction Ignore
-        $null = New-Item -Type Directory $templateDir
-        Copy-Item $repoDir/templates $templateDir
+    $staticDir = "$portalDir/01-portal/files/static"
+    Remove-Item -Recurse -Force $staticDir -ErrorAction Ignore
+    $null = New-Item -Type Directory $staticDir
+    Copy-Item -Recurse $buildDir/css $staticDir/css
+    Copy-Item -Recurse $repoDir/client/js $staticDir/js
 
-        Copy-Item $pigenDir/config-portal.sh $pigenBuildDir/config-portal
-
-        $configName = "portal"
-    }
-    "CameraOnly" {
-        prepareStage "agent"
-
-        $agentDir = "$pigenBuildDir/agent"
-
-        Copy-Item $rustBuildDir/lcsvc $agentDir/02-agent/files/lcagent
-
-        Copy-Item $pigenDir/config-agent.sh $pigenBuildDir/config-agent
-
-        $configName = "agent"
-    }
+    $stageList += " portal"
 }
+
+if ($Variant -ne "PortalOnly") {
+
+    prepareStage "stream"
+
+    $stageList += " stream"
+}
+
+prepareStage "finalize"
+$stageList += " finalize"
+
+$imgSuffix = switch ($Variant) {
+    "CameraOnly" { "-camera" }
+    "PortalOnly" { "-portal" }
+}
+@"
+export STAGE_LIST="$stageList"
+export LC_IMG_SUFFIX="$imgSuffix"
+"@ >> $pigenBuildDir/config
 
 Push-Location $pigenBuildDir
 Write-Host "building raspbian image"
 docker rm -v pigen_work
-./build-docker.sh -c config-$configName
+./build-docker.sh -c config
 Pop-Location
